add types and version sorting for the registry
Some of this is likely duplicated in the registry, so we may refactor
the types out into a shared package later.

Adds basic detector for registry module source strings. While this isn't
a thorough validation, this will eliminate anything that is definitely
not a registry module, and split out our host and module id strings.

lookupModuleVersions interrogates the registry for the available
versions of a particular module and the tree of dependencies.